Fix a bug caused by setTimeout circular reference
There is no easy way to clear previous setTimeout methods on the
updateSong method. Initial attempts at storing a reference on the
session object result in a circular reference due to the return
type of setTimeout. Instead, a polling-based appraoch will have to
be used on the client side to keep track of song progress.