Move the playNext function to the App component
The playNext function from the playbackControls function is moved
to app.jsx, so that it may be called when the current song has
finished playing (without need to manually select the next song).

Several changes had to be made to the function, as closures caused
the function to have stale references to both the trackQueue and
basePlaylist state variables. React's useRef method was used to
create references to the affected state, in order to ensure proper
functionality.

Additionally, the updateTrackInfo method was changed to now call
the playNext function on return from setTimeout, rather than
requesting the server for new track information.